Single p2e for both zero-shot and finetuning
